Refactored GameScreen to extract the game logic from the input/output logic
This commit aims to start the extraction of the game engine from the game specific implementation
